discurber/requirements.txt
astroid==1.6.5
backcall==0.1.0
decorator==4.3.0
Django==2.0.6
django-cors-headers==2.3.0
django-extensions==2.0.7
django-rest-framework==0.1.0
djangorestframework==3.8.2
ipdb==0.11
ipython==6.4.0
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.12.0
lazy-object-proxy==1.3.1
mccabe==0.6.1
parso==0.2.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
psycopg2==2.7.5
psycopg2-binary==2.7.5
ptyprocess==0.6.0
Pygments==2.2.0
pylint==1.9.2
pytz==2018.4
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
wrapt==1.10.11
